pyproject.toml
# Copyright 2023 Canonical Ltd.
# See LICENSE file for licensing details.

[tool.poetry]
package-mode = false
requires-poetry = ">=2.0.0"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
ops = "^2.17.0"
pydantic = "^1.10.7"
pyyaml = ">=6.0.1"
boto3 = "^1.36"
jsonschema = "^4.19.1"
tenacity = ">=8.0.1"
azure-storage-blob = "^12.24.0"

[tool.poetry.group.charm-libs.dependencies]
ops = "^2.17.0"
poetry-core = "*"
# grafana_agent/v0/cos_agent.py requires pydantic <2
# traefix_k8s/v2/ingress.py
pydantic = "^1.10.17, <2"
# grafana_agent/v0/cos_agent.py
# loki_k8s/v1/loki_push_api.py
# prometheus_k8s/v0/prometheus_scrape.py
pyyaml = "*"
# oathkeeper/v0/auth_proxy.py
jsonschema = "*"
# loki_k8s/v1/loki_push_api.py
# prometheus_k8s/v0/prometheus_scrape.py
cosl = "*"

[tool.poetry.group.format]
optional = true

[tool.poetry.group.format.dependencies]
ruff = ">=0.8.0"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
ruff = ">=0.8.0"
codespell = ">=2.2.2"
boto3-stubs = { extras = ["s3"], version = "^1.36" }

[tool.poetry.group.unit]
optional = true

[tool.poetry.group.unit.dependencies]
pytest = "^8.0.0"
coverage = { extras = ["toml"], version = ">7.0" }
ops = { version = "^2.17.0", extras = ["testing"] }
moto = { extras = ["s3"], version = "^5.0.22" }

[tool.poetry.group.integration]
optional = true

[tool.poetry.group.integration.dependencies]
pytest = "^8.0.0"
coverage = { extras = ["toml"], version = ">7.0" }
tenacity = "^9.0.0"
jubilant = "^1.0.1"

[tool.ruff]
line-length = 99
extend-exclude = ["__pycache__", "*.egg_info"]
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "C", "N", "D", "I001"]
extend-ignore = [
  "D203",
  "D204",
  "D213",
  "D215",
  "D400",
  "D401",
  "D404",
  "D406",
  "D407",
  "D408",
  "D409",
  "D413",
]
ignore = ["E501", "D107"]
per-file-ignores = { "tests/*" = ["D100", "D101", "D102", "D103", "D104", "E999"], "src/literals.py" = ["D101"] }
mccabe.max-complexity = 10

# Testing tools configuration
[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"
asyncio_mode = "auto"
markers = ["unstable"]

[tool.pyright]
include = ["src"]
extraPaths = ["./lib", "src"]
pythonVersion = "3.10"
pythonPlatform = "All"
